Consolidate RichTextEditor; constrain tiptap lingo
Jira ticket: CAMS-526
