--- user-interface/src/lib/components/cams/RichTextEditor/Editor.test.ts
@@ -1,757 +1,0 @@
-import { describe, expect, vi, beforeEach, afterEach, test } from 'vitest';
-import { Editor } from './Editor';
-import { MockSelectionService } from './SelectionService.humble';
-import { ZERO_WIDTH_SPACE } from '@/lib/components/cams/RichTextEditor/Editor.constants';
-import editorUtilities, { safelySetHtml } from './Editor.utilities';
-import { setCursorInParagraph, setCursorInElement } from './RichTextEditor.test-utils';
-
-const createPrintableKeyEvent = (key: string): React.KeyboardEvent<HTMLDivElement> =>
-  ({
-    key,
-    ctrlKey: false,
-    metaKey: false,
-    altKey: false,
-    preventDefault: vi.fn(),
-  }) as unknown as React.KeyboardEvent<HTMLDivElement>;
-
-const createNonPrintableKeyEvent = (): React.KeyboardEvent<HTMLDivElement> =>
-  ({
-    key: 'ArrowLeft',
-    ctrlKey: false,
-    metaKey: false,
-    altKey: false,
-    preventDefault: vi.fn(),
-  }) as unknown as React.KeyboardEvent<HTMLDivElement>;
-
-const createCtrlKeyEvent = (key: string): React.KeyboardEvent<HTMLDivElement> =>
-  ({
-    key,
-    ctrlKey: true,
-    metaKey: false,
-    altKey: false,
-    preventDefault: vi.fn(),
-  }) as unknown as React.KeyboardEvent<HTMLDivElement>;
-
-describe('Editor', () => {
-  let root: HTMLDivElement;
-  let editor: Editor;
-  let selectionService: MockSelectionService;
-
-  beforeEach(() => {
-    root = document.createElement('div');
-    document.body.appendChild(root);
-    selectionService = new MockSelectionService();
-    editor = new Editor(root, selectionService);
-  });
-
-  afterEach(() => {
-    document.body.removeChild(root);
-    vi.restoreAllMocks();
-  });
-
-  test('should be constructable', () => {
-    expect(editor).toBeInstanceOf(Editor);
-  });
-
-  describe('initialization', () => {
-    test('should initialize with an empty paragraph if the root is empty', () => {
-      expect(editorUtilities.safelyGetHtml(root)).toBe('<p>​</p>');
-    });
-
-    test('should not modify root if it already has content', () => {
-      const newRoot = document.createElement('div');
-      editorUtilities.safelySetHtml(newRoot, '<p>existing content</p>');
-      new Editor(newRoot, selectionService);
-      expect(newRoot.innerHTML).toBe('<p>existing content</p>');
-    });
-  });
-
-  describe('Editor: handleCtrlKey method', () => {
-    test.each([
-      { key: 'b', expectedFormat: 'strong', description: 'ctrl+b' },
-      { key: 'i', expectedFormat: 'em', description: 'ctrl+i' },
-      { key: 'u', expectedFormat: 'u', description: 'ctrl+u' },
-    ])(
-      'toggleSelection is called with the correct format when $description is pressed',
-      ({ key, expectedFormat }) => {
-        const event = createCtrlKeyEvent(key);
-        // @ts-expect-error - Accessing private property for testing
-        const toggleSelectionSpy = vi.spyOn(editor.formattingService, 'toggleSelection');
-        const result = editor.handleCtrlKey(event);
-
-        expect(result).toBe(true);
-        expect(event.preventDefault).toHaveBeenCalled();
-        expect(toggleSelectionSpy).toHaveBeenCalledWith(expectedFormat);
-      },
-    );
-
-    test('returns false for non-ctrl or non-cmd keys', () => {
-      const event = createCtrlKeyEvent('a');
-      const result = editor.handleCtrlKey(event);
-
-      expect(result).toBe(false);
-      expect(event.preventDefault).not.toHaveBeenCalled();
-    });
-  });
-});
-
-describe('Editor: handlePrintableKey method', () => {
-  let editor: Editor;
-  let container: HTMLDivElement;
-  let selectionService: MockSelectionService;
-
-  beforeEach(() => {
-    container = document.createElement('div');
-    document.body.appendChild(container);
-    selectionService = new MockSelectionService();
-    editor = new Editor(container, selectionService);
-  });
-
-  afterEach(() => {
-    document.body.removeChild(container);
-    vi.restoreAllMocks();
-  });
-
-  test('returns false for non-printable keys', () => {
-    const event = createNonPrintableKeyEvent();
-    const result = editor.handlePrintableKey(event);
-
-    expect(result).toBe(false);
-    expect(event.preventDefault).not.toHaveBeenCalled();
-  });
-
-  test('returns false for key combinations with modifiers', () => {
-    const event = createCtrlKeyEvent('a');
-    const result = editor.handlePrintableKey(event);
-
-    expect(result).toBe(false);
-    expect(event.preventDefault).not.toHaveBeenCalled();
-  });
-
-  test('returns false when no selection exists', () => {
-    selectionService.getCurrentSelection = vi.fn().mockReturnValue(null);
-
-    const event = createPrintableKeyEvent('a');
-    const result = editor.handlePrintableKey(event);
-
-    expect(result).toBe(false);
-    expect(event.preventDefault).not.toHaveBeenCalled();
-  });
-
-  test('allows typing in empty paragraph', () => {
-    // The editor starts with an empty paragraph containing zero-width space
-    const paragraph = container.querySelector('p');
-    expect(paragraph).not.toBeNull();
-    setCursorInParagraph(paragraph!, 1, selectionService);
-
-    const event = createPrintableKeyEvent('a');
-    const result = editor.handlePrintableKey(event);
-
-    expect(result).toBe(false); // Let browser handle naturally
-    expect(event.preventDefault).not.toHaveBeenCalled();
-  });
-
-  test('creates paragraph when typing directly in root', () => {
-    // Clear the container and position cursor in root
-    safelySetHtml(container, '');
-    const textNode = document.createTextNode('');
-    container.appendChild(textNode);
-
-    const range = selectionService.createRange();
-    range.setStart(textNode, 0);
-    range.collapse(true);
-    selectionService.setSelectionRange(range);
-
-    const event = createPrintableKeyEvent('a');
-    const result = editor.handlePrintableKey(event);
-
-    expect(result).toBe(true);
-    expect(event.preventDefault).toHaveBeenCalled();
-    expect(container.querySelector('p')).toBeTruthy();
-  });
-});
-
-// Tests for toggleSelection have been moved to FormattingService.test.ts
-
-describe('Editor: handleBackspaceOnEmptyContent', () => {
-  let editor: Editor;
-  let container: HTMLDivElement;
-  let selectionService: MockSelectionService;
-
-  const createBackspaceEvent = (): React.KeyboardEvent<HTMLDivElement> =>
-    ({
-      key: 'Backspace',
-      preventDefault: vi.fn(),
-    }) as unknown as React.KeyboardEvent<HTMLDivElement>;
-
-  const createOtherKeyEvent = (): React.KeyboardEvent<HTMLDivElement> =>
-    ({
-      key: 'Delete',
-      preventDefault: vi.fn(),
-    }) as unknown as React.KeyboardEvent<HTMLDivElement>;
-
-  beforeEach(() => {
-    container = document.createElement('div');
-    document.body.appendChild(container);
-    selectionService = new MockSelectionService();
-    editor = new Editor(container, selectionService);
-  });
-
-  afterEach(() => {
-    document.body.removeChild(container);
-    vi.restoreAllMocks();
-  });
-
-  test('returns false for non-Backspace keys', () => {
-    const event = createOtherKeyEvent();
-    const result = editor.handleBackspaceOnEmptyContent(event);
-
-    expect(result).toBe(false);
-    expect(event.preventDefault).not.toHaveBeenCalled();
-  });
-
-  test('returns false for null selection', () => {
-    selectionService.getCurrentSelection = vi.fn().mockReturnValue(null);
-
-    const event = createBackspaceEvent();
-    const result = editor.handleBackspaceOnEmptyContent(event);
-
-    expect(result).toBe(false);
-    expect(event.preventDefault).not.toHaveBeenCalled();
-  });
-
-  test('returns false for selection with no range', () => {
-    selectionService.getCurrentSelection = vi.fn().mockReturnValue({
-      rangeCount: 0,
-    });
-
-    const event = createBackspaceEvent();
-    const result = editor.handleBackspaceOnEmptyContent(event);
-
-    expect(result).toBe(false);
-    expect(event.preventDefault).not.toHaveBeenCalled();
-  });
-
-  test('returns false when findClosestAncestor returns null', () => {
-    const findClosestAncestorSpy = vi.spyOn(editorUtilities, 'findClosestAncestor');
-    findClosestAncestorSpy.mockReturnValue(null);
-
-    const event = createBackspaceEvent();
-    const result = editor.handleBackspaceOnEmptyContent(event);
-
-    expect(result).toBe(false);
-    expect(event.preventDefault).not.toHaveBeenCalled();
-  });
-
-  test('prevents deletion of last empty paragraph', () => {
-    // Editor starts with one empty paragraph
-    const paragraph = container.querySelector('p');
-    expect(paragraph).not.toBeNull();
-    setCursorInParagraph(paragraph!, 1, selectionService);
-
-    const event = createBackspaceEvent();
-    const result = editor.handleBackspaceOnEmptyContent(event);
-
-    expect(result).toBe(true);
-    expect(event.preventDefault).toHaveBeenCalled();
-    expect(container.querySelectorAll('p')).toHaveLength(1);
-  });
-
-  test('allows deletion when multiple paragraphs exist', () => {
-    safelySetHtml(container, `<p>First</p><p>${ZERO_WIDTH_SPACE}</p>`);
-    const emptyParagraph = container.querySelectorAll('p')[1];
-    setCursorInParagraph(emptyParagraph, 1, selectionService);
-
-    const event = createBackspaceEvent();
-    const result = editor.handleBackspaceOnEmptyContent(event);
-
-    expect(result).toBe(true);
-    expect(event.preventDefault).toHaveBeenCalled();
-    // Should merge with previous paragraph
-  });
-
-  test('handles case where previousSibling is not an element', () => {
-    // Create a text node as the previous sibling
-    const textNode = document.createTextNode('Text Node');
-    container.appendChild(textNode);
-    safelySetHtml(container, `${container.innerHTML}<p>${ZERO_WIDTH_SPACE}</p>`);
-
-    const emptyParagraph = container.querySelector('p');
-    expect(emptyParagraph).not.toBeNull();
-    setCursorInParagraph(emptyParagraph!, 1, selectionService);
-
-    const event = createBackspaceEvent();
-    const result = editor.handleBackspaceOnEmptyContent(event);
-
-    // Should not throw and should prevent default since we're handling it
-    expect(result).toBe(false);
-    expect(event.preventDefault).toHaveBeenCalled();
-  });
-
-  test('handles case where previousSibling has no firstChild', () => {
-    // Create an empty element as the previous sibling
-    const emptyElement = document.createElement('div');
-    container.appendChild(emptyElement);
-    safelySetHtml(container, `${container.innerHTML}<p>${ZERO_WIDTH_SPACE}</p>`);
-
-    const emptyParagraph = container.querySelector('p');
-    expect(emptyParagraph).not.toBeNull();
-    setCursorInParagraph(emptyParagraph!, 1, selectionService);
-
-    const event = createBackspaceEvent();
-    const result = editor.handleBackspaceOnEmptyContent(event);
-
-    // Should not throw and should prevent default since we're handling it
-    expect(result).toBe(false);
-    expect(event.preventDefault).toHaveBeenCalled();
-  });
-
-  test('handles backspace properly when previousSibling firstChild has insufficient length', () => {
-    safelySetHtml(container, '<p>First paragraph</p>');
-
-    const complexElement = document.createElement('p');
-    const emptySpan = document.createElement('span');
-    const textNode = document.createTextNode('Some content that makes textContent.length > 0');
-
-    complexElement.appendChild(emptySpan);
-    complexElement.appendChild(textNode);
-    container.appendChild(complexElement);
-
-    const zeroWidthParagraph = document.createElement('p');
-    zeroWidthParagraph.textContent = ZERO_WIDTH_SPACE;
-    container.appendChild(zeroWidthParagraph);
-
-    setCursorInParagraph(zeroWidthParagraph, 1, selectionService);
-
-    const event = createBackspaceEvent();
-
-    editor.handleBackspaceOnEmptyContent(event);
-
-    expect(event.preventDefault).toHaveBeenCalled();
-
-    expect(container.querySelector(`p[textContent="${ZERO_WIDTH_SPACE}"]`)).toBeFalsy();
-  });
-
-  test('allows normal backspace in non-empty paragraph', () => {
-    safelySetHtml(container, '<p>Hello world</p>');
-    const paragraph = container.querySelector('p');
-    expect(paragraph).not.toBeNull();
-    setCursorInParagraph(paragraph!, 5, selectionService);
-
-    const event = createBackspaceEvent();
-    const result = editor.handleBackspaceOnEmptyContent(event);
-
-    expect(result).toBe(false);
-    expect(event.preventDefault).not.toHaveBeenCalled();
-  });
-});
-
-describe('Editor: Additional coverage tests', () => {
-  let editor: Editor;
-  let container: HTMLDivElement;
-  let selectionService: MockSelectionService;
-
-  beforeEach(() => {
-    container = document.createElement('div');
-    document.body.appendChild(container);
-    selectionService = new MockSelectionService();
-    editor = new Editor(container, selectionService);
-  });
-
-  afterEach(() => {
-    document.body.removeChild(container);
-    vi.restoreAllMocks();
-  });
-
-  describe('toggleList - additional scenarios', () => {
-    test('toggleList handles nested lists correctly', () => {
-      safelySetHtml(container, '<ul><li>Item 1<ul><li>Nested item</li></ul></li></ul>');
-      const nestedItem = container.querySelector('ul ul li') as HTMLElement;
-      expect(nestedItem).not.toBeNull();
-      setCursorInElement(nestedItem, 0, selectionService);
-
-      editor.toggleList('ul');
-
-      // The nested list item should be converted to a paragraph
-      expect(container.innerHTML).toContain('<p>Nested item</p>');
-      // The parent list should still exist
-      expect(container.querySelector('ul')).toBeTruthy();
-    });
-
-    test('toggleList creates different list types', () => {
-      // Start with a paragraph
-      safelySetHtml(container, '<p>Item 1</p>');
-      const paragraph = container.querySelector('p') as HTMLElement;
-      expect(paragraph).not.toBeNull();
-      setCursorInElement(paragraph, 0, selectionService);
-
-      // Convert to unordered list
-      editor.toggleList('ul');
-
-      // Should now be an unordered list
-      expect(container.querySelector('ul')).toBeTruthy();
-      expect(container.querySelector('li')).toBeTruthy();
-      const li = container.querySelector('li');
-      expect(li).not.toBeNull();
-      expect(li!.textContent).toEqual(`${ZERO_WIDTH_SPACE}Item 1`);
-
-      // Clear and create a new paragraph
-      safelySetHtml(container, '<p>Item 1</p>');
-      const newParagraph = container.querySelector('p') as HTMLElement;
-      expect(newParagraph).not.toBeNull();
-      setCursorInElement(newParagraph, 0, selectionService);
-
-      // Convert to ordered list
-      editor.toggleList('ol');
-
-      // Should now be an ordered list
-      expect(container.querySelector('ol')).toBeTruthy();
-      expect(container.querySelector('li')).toBeTruthy();
-      const orderedLi = container.querySelector('li');
-      expect(orderedLi).not.toBeNull();
-      expect(orderedLi!.textContent).toEqual(`${ZERO_WIDTH_SPACE}Item 1`);
-    });
-
-    test('toggleList handles empty paragraphs correctly', () => {
-      safelySetHtml(container, `<p>${ZERO_WIDTH_SPACE}</p>`);
-      const paragraph = container.querySelector('p');
-      expect(paragraph).not.toBeNull();
-      setCursorInParagraph(paragraph!, 1, selectionService);
-
-      editor.toggleList('ul');
-
-      expect(container.querySelector('ul')).toBeTruthy();
-      expect(container.querySelector('li')).toBeTruthy();
-      expect(container.querySelector('p')).toBeFalsy();
-    });
-  });
-
-  test('handles cross-paragraph selection gracefully in toggleSelection', () => {
-    safelySetHtml(container, '<p>First paragraph</p><p>Second paragraph</p>');
-
-    const firstP = container.querySelectorAll('p')[0];
-    const secondP = container.querySelectorAll('p')[1];
-    const firstText = firstP.firstChild as Text;
-    const secondText = secondP.firstChild as Text;
-
-    const range = document.createRange();
-    range.setStart(firstText, 5);
-    range.setEnd(secondText, 6);
-    selectionService.setSelectionRange(range);
-
-    // Should not apply formatting across paragraphs
-    editor.toggleSelection('strong');
-
-    // Content should remain unchanged
-    expect(container.innerHTML).toBe('<p>First paragraph</p><p>Second paragraph</p>');
-  });
-
-  test('properly merges adjacent similar elements after formatting', () => {
-    safelySetHtml(container, '<p><strong>Bold</strong> text <strong>more</strong></p>');
-
-    const pElement = container.querySelector('p')!;
-    const textNode = pElement.childNodes[1] as Text;
-
-    const range = document.createRange();
-    range.setStart(textNode, 1);
-    range.setEnd(textNode, 5);
-    selectionService.setSelectionRange(range);
-
-    editor.toggleSelection('strong');
-
-    // The normalization should happen automatically, but let's check what actually happened
-    // The expected behavior is that all adjacent strong elements get merged
-    expect(container.innerHTML).toContain('<strong>Bold</strong>');
-    expect(container.innerHTML).toContain('<strong>text</strong>');
-    expect(container.innerHTML).toContain('<strong>more</strong>');
-  });
-
-  test('initializes with empty paragraph when container is empty', () => {
-    const newContainer = document.createElement('div');
-    new Editor(newContainer, selectionService);
-    expect(newContainer.innerHTML).toBe(`<p>${ZERO_WIDTH_SPACE}</p>`);
-  });
-
-  test('toggles formatting with collapsed cursor', () => {
-    safelySetHtml(container, `<p>${ZERO_WIDTH_SPACE}</p>`);
-    setCursorInParagraph(container.querySelector('p')!, 1, selectionService);
-
-    editor.toggleSelection('strong');
-
-    expect(container.innerHTML).toContain('<strong>');
-    expect(container.innerHTML).toContain(ZERO_WIDTH_SPACE);
-  });
-
-  test('list splitting behavior when toggling middle item', () => {
-    editorUtilities.safelySetHtml(
-      container,
-      '<ul><li>Item 1</li><li>Item 2</li><li>Item 3</li></ul>',
-    );
-
-    const middleLi = container.querySelectorAll('li')[1];
-    const range = document.createRange();
-    range.selectNodeContents(middleLi);
-    selectionService.setSelectionRange(range);
-
-    editor.toggleList('ul');
-
-    const html = editorUtilities.safelyGetHtml(container);
-    expect(html).toMatch(/<ul><li>Item 1<\/li><\/ul>/);
-    expect(html).toMatch(/<p>Item 2<\/p>/);
-    expect(html).toMatch(/<ul><li>Item 3<\/li><\/ul>/);
-  });
-
-  test('single item list converts to paragraph', () => {
-    safelySetHtml(container, '<ul><li>Only Item</li></ul>');
-
-    const firstLi = container.querySelector('li');
-    const range = document.createRange();
-    range.selectNodeContents(firstLi!);
-    selectionService.setSelectionRange(range);
-
-    editor.toggleList('ul');
-
-    expect(container.innerHTML).toBe('<p>Only Item</p>');
-  });
-
-  test('preserves nested formatting when toggling off one format', () => {
-    safelySetHtml(container, `<p><em><strong>${ZERO_WIDTH_SPACE}</strong></em></p>`);
-    const strongElement = container.querySelector('strong')!;
-    setCursorInElement(strongElement, 1, selectionService);
-
-    editor.toggleSelection('strong');
-
-    // The implementation creates a new structure when exiting formatting
-    // It preserves the italic formatting and positions the cursor appropriately
-    expect(container.innerHTML).toContain('<em>');
-    // The strong element might still exist in the structure but cursor has moved out
-    expect(container.innerHTML).toMatch(/<p><em>.*<\/em><\/p>/);
-  });
-
-  describe('complex interactions between methods', () => {
-    test('formatting and list conversion', () => {
-      // Create a paragraph with text
-      safelySetHtml(container, '<p>Test paragraph</p>');
-
-      // Apply formatting to the paragraph
-      const paragraph = container.querySelector('p')!;
-      const range = selectionService.createRange();
-      range.setStart(paragraph.firstChild!, 0);
-      range.setEnd(paragraph.firstChild!, 4); // Select "Test"
-      selectionService.setSelectionRange(range);
-
-      editor.toggleSelection('strong');
-
-      // Verify formatting was applied
-      expect(container.innerHTML).toContain('<strong>Test</strong>');
-
-      // Create a new paragraph for list conversion
-      safelySetHtml(container, '<p>List item</p>');
-      const newParagraph = container.querySelector('p')!;
-      setCursorInParagraph(newParagraph, 0, selectionService);
-
-      // Convert to list
-      editor.toggleList('ul');
-
-      // Verify list was created
-      expect(container.querySelector('ul')).toBeTruthy();
-      expect(container.querySelector('li')).toBeTruthy();
-      expect(container.querySelector('li')!.textContent).toEqual(`${ZERO_WIDTH_SPACE}List item`);
-    });
-
-    test('convertParagraphToList calls setCursorInListItem', () => {
-      // Create a paragraph with text
-      safelySetHtml(container, '<p>Test paragraph</p>');
-      const paragraph = container.querySelector('p');
-      expect(paragraph).not.toBeNull();
-      setCursorInParagraph(paragraph!, 0, selectionService);
-
-      // Convert to list
-      editor.toggleList('ul');
-
-      // Verify list was created
-      expect(container.querySelector('ul')).toBeTruthy();
-      expect(container.querySelector('li')).toBeTruthy();
-    });
-
-    test('setCursorInListItem is called during list conversion', () => {
-      // Create a paragraph with text
-      safelySetHtml(container, '<p>Test paragraph</p>');
-      const paragraph = container.querySelector('p');
-      expect(paragraph).not.toBeNull();
-      setCursorInParagraph(paragraph!, 0, selectionService);
-
-      // Convert to list
-      editor.toggleList('ul');
-
-      // Verify list was created
-      expect(container.querySelector('ul')).toBeTruthy();
-      expect(container.querySelector('li')).toBeTruthy();
-    });
-
-    test('convertParagraphToList returns early when no selection exists', () => {
-      // Create a paragraph
-      safelySetHtml(container, '<p>Test paragraph</p>');
-
-      // Mock the getCurrentSelection to return null
-      const originalGetSelection = selectionService.getCurrentSelection;
-      selectionService.getCurrentSelection = vi.fn().mockReturnValue(null);
-
-      // Try to convert to list
-      editor.toggleList('ul');
-
-      // Verify list was not created
-      expect(container.querySelector('ul')).toBeFalsy();
-      expect(container.innerHTML).toBe('<p>Test paragraph</p>');
-
-      // Restore the original method
-      selectionService.getCurrentSelection = originalGetSelection;
-    });
-
-    test('list conversion followed by formatting', () => {
-      safelySetHtml(container, '<p>Test paragraph</p>');
-      const paragraph = container.querySelector('p');
-      expect(paragraph).not.toBeNull();
-      setCursorInParagraph(paragraph!, 5, selectionService);
-
-      // Convert to list first
-      editor.toggleList('ul');
-
-      // Then apply formatting
-      const listItem = container.querySelector('li');
-      expect(listItem).not.toBeNull();
-      const range = selectionService.createRange();
-      range.setStart(listItem!.childNodes[1]!, 0);
-      range.setEnd(listItem!.childNodes[1]!, 4); // Select "Test"
-      selectionService.setSelectionRange(range);
-      editor.toggleSelection('em');
-
-      // Should have formatting inside list
-      expect(container.querySelector('ul')).toBeTruthy();
-      expect(container.querySelector('li')).toBeTruthy();
-      expect(container.innerHTML).toContain('<em>Test</em>');
-    });
-
-    test('key handling interactions', () => {
-      safelySetHtml(container, '<p>Test</p>');
-      const paragraph = container.querySelector('p');
-      expect(paragraph).not.toBeNull();
-      setCursorInParagraph(paragraph!, 4, selectionService);
-
-      // Create Enter key event
-      const enterEvent = {
-        key: 'Enter',
-        preventDefault: vi.fn(),
-      } as unknown as React.KeyboardEvent<HTMLDivElement>;
-
-      // Handle Enter key to create new paragraph
-      editor.handleEnterKey(enterEvent);
-
-      // Now we should have two paragraphs
-      expect(container.querySelectorAll('p')).toHaveLength(2);
-
-      // Create Backspace key event
-      const backspaceEvent = {
-        key: 'Backspace',
-        preventDefault: vi.fn(),
-      } as unknown as React.KeyboardEvent<HTMLDivElement>;
-
-      // Position cursor at start of second paragraph
-      const secondParagraph = container.querySelectorAll('p')[1];
-      setCursorInParagraph(secondParagraph, 0, selectionService);
-
-      // Handle Backspace to delete the empty paragraph
-      editor.handleBackspaceOnEmptyContent(backspaceEvent);
-
-      // Should merge paragraphs or remove the empty one
-      expect(container.querySelectorAll('p')).toHaveLength(1);
-    });
-
-    test('multiple formatting operations', () => {
-      safelySetHtml(container, '<p>Test paragraph</p>');
-      const paragraph = container.querySelector('p');
-      expect(paragraph).not.toBeNull();
-      const range = selectionService.createRange();
-      range.setStart(paragraph!.firstChild!, 0);
-      range.setEnd(paragraph!.firstChild!, 4); // Select "Test"
-      selectionService.setSelectionRange(range);
-
-      // Apply bold formatting
-      editor.toggleSelection('strong');
-      expect(container.innerHTML).toContain('<strong>');
-
-      // Apply italic formatting to the same text
-      // First, reselect the text inside the strong element
-      const strongElement = container.querySelector('strong');
-      expect(strongElement).not.toBeNull();
-      const newRange = selectionService.createRange();
-      newRange.selectNodeContents(strongElement!);
-      selectionService.setSelectionRange(newRange);
-
-      editor.toggleSelection('em');
-
-      // Verify both formats are applied (the order might vary)
-      const html = container.innerHTML;
-      expect(html.includes('<strong><em>') || html.includes('<em><strong>')).toBe(true);
-
-      // Text should remain unchanged
-      expect(container.textContent).toContain('Test paragraph');
-    });
-
-    test('applying and removing a single format', () => {
-      safelySetHtml(container, '<p>Test paragraph</p>');
-      const paragraph = container.querySelector('p');
-      expect(paragraph).not.toBeNull();
-      const range = selectionService.createRange();
-      range.setStart(paragraph!.firstChild!, 0);
-      range.setEnd(paragraph!.firstChild!, 4); // Select "Test"
-      selectionService.setSelectionRange(range);
-
-      // Apply bold formatting
-      editor.toggleSelection('strong');
-      expect(container.innerHTML).toContain('<strong>');
-
-      // Remove bold formatting
-      const strongElement = container.querySelector('strong');
-      expect(strongElement).not.toBeNull();
-      const newRange = selectionService.createRange();
-      newRange.selectNodeContents(strongElement!);
-      selectionService.setSelectionRange(newRange);
-
-      editor.toggleSelection('strong');
-      expect(container.innerHTML).not.toContain('<strong>');
-
-      // Text should remain unchanged
-      expect(container.textContent).toContain('Test paragraph');
-    });
-  });
-  describe('complex DOM structures in text handling', () => {
-    test('handles non-text node firstChild in paragraphs', () => {
-      // Create a paragraph with a non-text firstChild (a span)
-      safelySetHtml(container, '<p><span>Text in span</span></p>');
-
-      // Get the paragraph
-      const paragraph = container.querySelector('p');
-      expect(paragraph).not.toBeNull();
-
-      // Set cursor in the paragraph
-      const range = selectionService.createRange();
-      const { firstChild } = paragraph!;
-      expect(firstChild).not.toBeNull();
-      const firstChildOfFirstChild = firstChild!.firstChild;
-      expect(firstChildOfFirstChild).not.toBeNull();
-      range.setStart(firstChildOfFirstChild!, 0);
-      range.collapse(true);
-      selectionService.setSelectionRange(range);
-
-      // Trigger a method that uses the textNode extraction logic
-      editor.handleEnterKey({
-        key: 'Enter',
-        preventDefault: vi.fn(),
-      } as unknown as React.KeyboardEvent<HTMLDivElement>);
-
-      // Verify the operation completed without errors
-      expect(container.querySelectorAll('p').length).toBeGreaterThan(1);
-    });
-  });
-});
