--- user-interface/src/lib/components/cams/RichTextEditor/RichTextButton.scss
@@ -1,9 +1,0 @@
-.usa-button.rich-text-button {
-  width: 40px;
-  height: 40px;
-  .usa-icon {
-    width: 40px;
-    height: 40px;
-    display: block;
-  }
-}
