+++ user-interface/src/lib/components/cams/RichTextEditor/RichTextIcon.tsx
@@ -1,17 +1,101 @@
 import Icon from '@/lib/components/uswds/Icon';
 
-export interface RichTextIconProps {
-  name: string;
+export function NumberedListIcon() {
+  return (
+    <svg
+      fill="#ffffff"
+      width="1rem"
+      height="1rem"
+      version="1.1"
+      id="Capa_1"
+      xmlns="http://www.w3.org/2000/svg"
+      xmlnsXlink="http://www.w3.org/1999/xlink"
+      viewBox="0 0 94.479 94.479"
+      xmlSpace="preserve"
+    >
+      <g>
+        <g>
+          <path
+            d="M92.479,6.821H26.823c-1.104,0-2,0.896-2,2V19.58c0,1.104,0.896,2,2,2h65.656c1.104,0,2-0.896,2-2V8.821
+			C94.479,7.717,93.584,6.821,92.479,6.821z"
+          />
+          <path
+            d="M92.479,40.457H26.823c-1.104,0-2,0.896-2,2v10.761c0,1.104,0.896,2,2,2h65.656c1.104,0,2-0.896,2-2V42.457
+			C94.479,41.352,93.584,40.457,92.479,40.457z"
+          />
+          <path
+            d="M92.479,75.109H26.823c-1.104,0-2,0.896-2,2v10.759c0,1.104,0.896,2,2,2h65.656c1.104,0,2-0.896,2-2V77.109
+			C94.479,76.004,93.584,75.109,92.479,75.109z"
+          />
+          <path
+            d="M1.842,11.104c0.423,0.051,0.848-0.08,1.167-0.36l0.914-0.799v12.973c0,0.828,0.672,1.5,1.5,1.5h1.546
+			c0.828,0,1.5-0.672,1.5-1.5V5.567c0-0.828-0.672-1.5-1.5-1.5H5.378c-0.356,0-0.701,0.127-0.973,0.358L0.527,7.728
+			c-0.589,0.501-0.7,1.369-0.256,2.003l0.522,0.745C1.04,10.825,1.42,11.053,1.842,11.104z"
+          />
+          <path
+            d="M2.546,37.5c-0.647,0.23-1.056,0.872-0.99,1.556l0.096,0.999c0.048,0.497,0.339,0.938,0.778,1.176
+			c0.438,0.238,0.967,0.243,1.409,0.011c1.18-0.615,1.923-0.775,2.922-0.534c0.71,0.163,1.188,0.586,1.438,1.225
+			c0.378,1.143,0.395,2.957-2.854,6.873l-2.506,2.996c-0.589,0.688-0.99,1.119-1.302,1.404c-0.312,0.284-0.488,0.688-0.488,1.107
+			v1.475c0,0.828,0.672,1.5,1.5,1.5h9.132c0.828,0,1.5-0.672,1.5-1.5V55.01c0-0.828-0.672-1.5-1.5-1.5H6.896
+			c0.458-0.515,1.037-1.211,1.716-2.063c2.803-3.468,5.061-6.645,4.292-10.376c-0.309-1.672-1.247-2.966-2.731-3.752
+			c-0.826-0.424-1.879-0.639-3.127-0.639C5.629,36.68,4.031,36.971,2.546,37.5z"
+          />
+          <path
+            d="M10.491,79.516c0.25-0.135,0.487-0.289,0.74-0.483c1.718-1.313,1.801-3.335,1.786-3.885
+			c-0.001-1.878-0.712-3.318-2.166-4.401c-1.026-0.77-2.534-1.177-4.361-1.177c-1.347,0-2.787,0.212-4.057,0.596
+			c-0.668,0.202-1.107,0.839-1.063,1.533l0.058,0.88c0.031,0.471,0.28,0.896,0.673,1.155c0.394,0.259,0.886,0.319,1.327,0.159
+			c0.864-0.307,1.571-0.45,2.223-0.45c0.229,0,0.457,0.019,0.722,0.063c0.94,0.127,1.563,0.549,1.9,1.283
+			c0.237,0.52,0.19,1.445-0.083,1.885c-0.654,1.014-2.438,1.217-3.443,1.25c-0.81,0.024-1.453,0.688-1.453,1.498v0.658
+			c0,0.828,0.687,1.5,1.515,1.5c1.575,0,2.311,0.309,2.596,0.485c0.737,0.465,1.065,1.043,1.066,1.918
+			c0.025,0.912-0.348,1.606-1.129,2.121c-1.014,0.651-2.64,0.647-4.505,0.032c-0.438-0.146-0.918-0.082-1.302,0.175
+			c-0.384,0.255-0.629,0.672-0.665,1.132l-0.068,0.872C0.75,88.973,1.135,89.588,1.75,89.83c0.33,0.13,0.551,0.183,1.022,0.281
+			c0.857,0.196,1.84,0.302,2.842,0.302c1.493,0,4.248-0.247,5.944-1.919c1.657-1.69,1.974-4.11,1.56-5.729
+			C12.636,80.704,11.461,79.946,10.491,79.516z"
+          />
+        </g>
+      </g>
+    </svg>
+  );
 }
 
-export function RichTextIcon(props: RichTextIconProps) {
-  const { name } = props;
-
-  return iconMap[name] ?? '';
+export function BulletListIcon() {
+  return (
+    <svg
+      fill="#ffffff"
+      width="1rem"
+      height="1rem"
+      version="1.1"
+      id="Capa_1"
+      xmlns="http://www.w3.org/2000/svg"
+      xmlnsXlink="http://www.w3.org/1999/xlink"
+      viewBox="0 0 95.494 95.494"
+      xmlSpace="preserve"
+    >
+      <g>
+        <g>
+          <circle cx="10.906" cy="17.294" r="10.906" />
+          <circle cx="10.906" cy="47.294" r="10.906" />
+          <circle cx="10.906" cy="78.2" r="10.906" />
+          <g>
+            <path
+              d="M34.258,24.43h59.236c0.529,0,1.039-0.211,1.414-0.586s0.586-0.884,0.586-1.414l-0.002-10.271c0-1.104-0.896-2-2-2
+				H34.258c-1.104,0-2,0.896-2,2V22.43C32.258,23.534,33.153,24.43,34.258,24.43z"
+            />
+            <path
+              d="M93.492,40.158H34.258c-1.104,0-2,0.896-2,2v10.271c0,1.104,0.896,2,2,2h59.236c0.529,0,1.039-0.211,1.414-0.586
+				s0.586-0.885,0.586-1.414l-0.002-10.271C95.492,41.054,94.598,40.158,93.492,40.158z"
+            />
+            <path
+              d="M93.492,71.064H34.258c-1.104,0-2,0.896-2,2v10.271c0,1.104,0.896,2,2,2h59.236c0.529,0,1.039-0.211,1.414-0.586
+				s0.586-0.885,0.586-1.414l-0.002-10.271C95.492,71.96,94.598,71.064,93.492,71.064z"
+            />
+          </g>
+        </g>
+      </g>
+    </svg>
+  );
 }
 
-const iconMap: Record<string, JSX.Element> = {
-  'numbered-list': <img src="/numbered-list.svg" alt="numbered list icon" />,
-  'bulleted-list': <img src="/bullet-list.svg" alt="bulleted list icon" />,
-  link: <Icon name="link"></Icon>,
-};
+export function LinkIcon() {
+  return <Icon name="link"></Icon>;
+}
