+++ user-interface/src/lib/components/cams/RichTextEditor/implementation-plan.md
