--- user-interface/src/lib/components/cams/TiptapEditor/TiptapEditor.scss
@@ -1,154 +1,0 @@
-.tiptap-editor-container {
-  .tiptap-editor-wrapper {
-    border: 1px solid #ccc; // match RichTextEditor border
-    border-radius: 0.25rem;
-    min-height: 120px;
-
-    .tiptap-editor-toolbar {
-      position: relative;
-      display: flex;
-      gap: 0.25rem;
-      padding: 0.5rem;
-      border-bottom: 1px solid #e0e0e0;
-      //background-color: #f8f9fa;
-      background-color: #ffffff;
-      flex-wrap: wrap;
-
-      button {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        font-weight: bold;
-        width: 2.5rem;
-        height: 2.5rem;
-        padding: 0 0.5rem;
-        margin-right: 0;
-        border: none;
-        border-radius: 0.25rem;
-        background-color: transparent;
-        color: #1b1b1b;
-
-        svg {
-          width: 1.2rem;
-          height: 1.2rem;
-          fill: #1b1b1b;
-        }
-
-        &:nth-child(2) {
-          font-style: italic;
-        }
-
-        &:nth-child(3) {
-          text-decoration: underline;
-        }
-      }
-      button:hover {
-        background-color: #e9ecef;
-        color: #1b1b1b;
-        cursor: pointer;
-      }
-      button.is-active {
-        background-color: #71767a;
-        color: #ffffff;
-        svg {
-          fill: #ffffff;
-        }
-      }
-
-      .tiptap-link-popover {
-        position: absolute;
-        display: flex;
-        align-items: center;
-        top: 3.5rem;
-        left: 1rem;
-        z-index: 10;
-        background-color: #f0f0f0;
-        border: 1px solid #ccc;
-        border-radius: 0.25rem;
-        padding: 0.5rem;
-        box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-        width: calc(100% - 2rem);
-
-        .display-text-input {
-          margin-left: 0.5rem;
-        }
-
-        input {
-          width: 50%;
-        }
-
-        button {
-          display: inline;
-        }
-        button:hover {
-          background-color: #005ea2;
-          svg {
-            fill: #ffffff;
-          }
-        }
-      }
-
-      .tiptap-link-apply, .tiptap-link-delete {
-        border: none;
-        border-radius: 0.25rem;
-        padding: 0.1rem 0.1rem;
-        margin-left: 0.5rem;
-        width: 1.5rem;
-        height: 1.5rem;
-
-        &:hover {
-          color: #005ea6;
-          cursor: pointer;
-        }
-      }
-    }
-
-    .tiptap-editor {
-      height: 12rem;
-      outline: none;
-      display: flex;
-
-      &.disabled {
-        background-color: #f0f0f0;
-        color: #565c65;
-        cursor: not-allowed;
-      }
-
-      .tiptap {
-        flex-basis: 100%;
-        padding: 0.75rem;
-      }
-
-      // Basic content styling
-      p {
-        margin: 0 0 0.5rem 0;
-
-        &:last-child {
-          margin-bottom: 0;
-        }
-      }
-
-      ul,
-      ol {
-        margin: 0.5rem 0;
-        padding-left: 1.5rem;
-      }
-
-      li {
-        margin: 0.25rem 0;
-      }
-
-      strong {
-        font-weight: bold;
-      }
-
-      em {
-        font-style: italic;
-      }
-
-      u {
-        text-decoration: underline;
-      }
-    }
-  }
-}
