--- user-interface/src/lib/components/cams/TiptapEditor/index.ts
@@ -1,2 +1,0 @@
-export { default as TiptapEditor } from './TiptapEditor';
-export type { TiptapEditorRef, TiptapEditorProps } from './TiptapEditor';
